[INLONG-8296][Manager] Move LoginUserUtils to pojo to avoid cycle dependency
