+++ inlong-manager/manager-web/src/main/java/org/apache/inlong/manager/web/auth/openapi/OpenAPIFilter.java
@@ -19,7 +19,7 @@ package org.apache.inlong.manager.web.auth.openapi;
 
 import org.apache.inlong.common.util.BasicAuth;
 import org.apache.inlong.manager.pojo.user.UserInfo;
-import org.apache.inlong.manager.service.user.LoginUserUtils;
+import org.apache.inlong.manager.pojo.user.LoginUserUtils;
 
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
